+++ src/IDBTransaction.js
@@ -213,7 +213,6 @@ IDBTransaction.prototype.objectStore = function (objectStoreName) {
     if (arguments.length === 0) {
         throw new TypeError('No object store name was specified');
     }
-    objectStoreName = util.stripNUL(objectStoreName);
     if (!this.__active) {
         throw createDOMException('InvalidStateError', 'A request was placed against a transaction which is currently not active, or which is finished');
     }